:bug: API test fixes
