+++ docs/Pokedex/PokedexEntries.yaml
@@ -35,6 +35,19 @@ get:
                                 type: array
                                 items:
                                     $ref: './Schemas/PokedexEntryListModel.yaml'
+        default:
+            description: Error
+            content:
+                application/vnd.api+json:
+                    schema:
+                        type: object
+                        required:
+                            - data
+                        properties:
+                            data:
+                                type: array
+                                items:
+                                    $ref: '../Common/Schemas/Error.yaml'
 post:
     summary: 'Create Pokedex entry'
     operationId: createPokedexEntry
